cne_mvp_v4/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-multipart==0.0.9
pandas==2.2.2
openpyxl==3.1.2
pdfplumber==0.11.0
python-docx==1.1.2
rapidfuzz==3.6.2
Jinja2==3.1.4
orjson==3.10.6
spacy==3.7.4
# Se for usar GPU: cupy-cudaXXX e torch CUDA via wheels locais (opcional)
